// packages/draw/editor-new/src/core/types.ts
export interface BaseEditor {
    enable: boolean
}
export interface Point {
    x: number
    y: number
}
export interface DragState {
    clickPoint: Point
    movePoint: Point
    originPosition: RectPositioning
    originConfigPosition: RectPositioning
    clickDeltaPoint: Point
}
export interface HoverState {
    hoveredBlockId: string
}
export enum EditorMode {
    None = 'none',
    Arrow = 'arrow',
    Select = 'select',
}

export interface ArrowState {
    start?: string
    startPosition?: Point
    end?: string
    endPosition?: Point
}
export interface SelectEvent {
    blockId: string
    x: number
    y: number
}
export interface EnabledEditorState {
    drag?: DragState
    hover?: HoverState | null
    selectedNodeIds?: string[]
    pins: PinState
    mode?: EditorMode
    arrowState?: ArrowState | null
    onSelect?: (select: SelectEvent) => void
    isViewMode?: boolean
    openLayoutSetting?: string
}

export interface PinState {
    selectedPinType?: PinType | null
}

export type SvgRef = Ref<SVGSVGElement | null>;
export type SvgGRef = Ref<SVGGElement | null>;

export interface MousePaletteState {
    pressed: boolean
    blockName?: string
}
export interface MouseState {
    position: Point
    pressed: boolean
    pressedPosition: Point
    palette: MousePaletteState
    leave: boolean
}
export interface HistoryChange {
    type: 'changes'
    changes: ChangedItem[]
}
export interface History {
    items: HistoryChange[]
    current?: number
}
export interface ZoomState {
    transform: string
    translate: { x: number; y: number }
    scale: number
}
export interface EnabledEditor extends BaseEditor {
    enable: true
    id?: string
    wrap: (node: any) => any
    state: EnabledEditorState
    showDebugWindow: boolean
    zoomState: ZoomState
    svg: SVGSVGElement | null
    workEl: SVGSVGElement | SVGGElement | null
    mouseState: MouseState
    config: ArchitectureDescription
    initialConfig: ArchitectureDescription
    architecture: Architecture
    makeChange: (change: Change) => void
    component: any
    history: History
    assets: AssetConfig
}
export interface DisabledEditor extends BaseEditor {
    enable: false
}

export type Editor = EnabledEditor | DisabledEditor;